+++ actions/acl_clone.py
@@ -1,11 +1,11 @@
 from lib import action
 
 
-class ConsulAclDestroyAction(action.ConsulBaseAction):
+class ConsulAclCloneAction(action.ConsulBaseAction):
     def run(self, acl_id, consul_profile=None, token=None):
         self._create_client(consul_profile)
         # Action parameter "token" defaults to an empty string when no input is provided.
         # Consul-Python tests for None to use the agent's token, so "token" set to None here.
         if token == "":
             token = None
-        return (True, self.consul.acl.destroy(acl_id, token=token))
+        return (True, self.consul.acl.clone(acl_id, token=token))