+++ actions/agent_service_deregister.py
@@ -2,5 +2,6 @@ from lib import action
 
 
 class ConsulServiceDeregisterAction(action.ConsulBaseAction):
-    def run(self, service_id):
+    def run(self, service_id, consul_profile=None):
+        self._create_client(consul_profile)
         return (True, self.consul.agent.service.deregister(service_id=service_id))